x86/platform/amd: Move the <asm/amd-ibs.h> header to <asm/amd/ibs.h>
commit 3846389c03a8518884f09056611619bd1461ffc7 upstream.

Collect AMD specific platform header files in <asm/amd/*.h>.

Acked-by: Borislav Petkov (AMD) <bp@alien8.de>
Cc: H. Peter Anvin <hpa@zytor.com>
Cc: Linus Torvalds <torvalds@linux-foundation.org>
Cc: Mario Limonciello <superm1@kernel.org>
Link: https://lore.kernel.org/r/20250413084144.3746608-2-mingo@kernel.org
